add faer as an alternative eigenvalue backend
This adds the ability to select different backends for the eigenvalue
computation. Each backend is optional and must be enabled with a feature
flag. The backend to use can be selected with the funtion
pm_remez_with_backend. The regular pm_remez function uses the default
backend, which depends on which feature flags are enabled.

faer is added as an alternative eigenvalue backend, in addition to
the existing LAPACK backend through ndarray-linalg.

Signed-off-by: Daniel Estévez <daniel@destevez.net>